requirements.txt
fastapi>=0.111
uvicorn>=0.30
redis>=5.0
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
